# openapi.json
{
"openapi": "3.0.0",
"info": {
"title": "Task API",
"description": "A simple Express API for managing tasks.",
"version": "1.0.0"
},
"servers": [
{
"url": "http://localhost:8000",
"description": "Local development server"
}
],
"paths": {
"/": {
"get": {
"summary": "API Information",
"description": "Returns basic information about the API and available endpoints.",
"responses": {
"200": {
"description": "Successful response",
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"name": { "type": "string" },
"version": { "type": "string" },
"endpoints": {
"type": "array",
"items": { "type": "string" }
}
}
}
}
}
}
}
}
},
"/health": {
"get": {
"summary": "Health Check",
"description": "Returns the health status of the server.",
"responses": {
"200": {
"description": "Successful response",
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"status": { "type": "string", "example": "okay" }
}
}
}
}
}
}
}
},
"/tasks": {
"get": {
"summary": "List all tasks",
"description": "Retrieves a list of tasks.",
"responses": {
"200": {
"description": "A list of tasks",
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"tasks": {
"type": "array",
"items": {
"$ref": "#/components/schemas/Task"
}
}
}
}
}
}
}
}
},
"post": {
"summary": "Create a new task",
"description": "Creates a new task with the provided title. Tasks are created with 'done' set to false by default.",
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"title": { "type": "string", "example": "Wash the car" }
},
"required": ["title"]
}
}
}
},
"responses": {
"201": {
"description": "Task successfully created",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Task"
}
}
}
},
"400": {
"description": "Invalid request body (e.g., missing title)",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Error"
}
}
}
}
}
}
},
"/tasks/{id}": {
"parameters": [
{
"name": "id",
"in": "path",
"description": "The unique ID of the task",
"required": true,
"schema": {
"type": "integer"
}
}
],
"get": {
"summary": "Get a specific task",
"description": "Retrieves a single task by its ID.",
"responses": {
"200": {
"description": "Task found",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Task"
}
}
}
},
"404": {
"description": "Task not found",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Error"
}
}
}
}
}
},
"put": {
"summary": "Update a task",
"description": "Updates a task's title, completion status, or both.",
"requestBody": {
"required": true,
"content": {
"application/json": {
"schema": {
"type": "object",
"properties": {
"title": { "type": "string", "example": "Wash the car thoroughly" },
"done": { "type": "boolean", "example": true }
}
}
}
}
},
"responses": {
"200": {
"description": "Task successfully updated",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Task"
}
}
}
},
"400": {
"description": "Invalid update parameters",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Error"
}
}
}
},
"404": {
"description": "Task not found",
"content": {
"application/json": {
"schema": {
"$ref": "#/components/schemas/Error"
}
}
}
}
}
}
}
},
"components": {
"schemas": {
"Task": {
"type": "object",
"properties": {
"id": {
"type": "integer",
"example": 1
},
"title": {
"type": "string",
"example": "read a book"
},
"done": {
"type": "boolean",
"example": false
}
}
},
"Error": {
"type": "object",
"properties": {
"error": {
"type": "string",
"example": "Task not found"
}
}
}
}
}
}